fix postman tests

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -54,10 +54,16 @@ public class FilmService {
     }
 
     public void addLike(Long filmID, Long userID) {
+        if (filmStorage.getFilm(filmID) == null) {
+            throw new NotFoundException("Не найдено");
+        }
         likeStorage.deleteLike(userID, filmID);
     }
 
     public void removeLike(Long filmID, Long userID) {
+        if (filmStorage.getFilm(filmID) == null) {
+            throw new NotFoundException("Не найдено");
+        }
         likeStorage.deleteLike(userID, filmID);
     }
